"No Mobile Support" Page & More Correct Menu Open/Close Logic
Created:
	- No Mobile Support Page
	- Error Header/Footer Templates
Implemented Portrait Menu State Machine
  Default behavior is now 'none' for transitions except for open/close
of the button. This is not only simpler and more in line with the
functionality we are trying to achieve, but Safari (of course) tries to
transition the header height on orientation changes before the css can
be applied, in behavior that is always opposite of Chrome or Firefox.