Split 'About' -> 'About & Design'
Replaced 'Projects' card with 'Design' after splitting 'About Me' into a
more brief about page and a page dedicated to my thoughts on high level
software design.
Templatized 'Articles' to reduce overhead of creating new pages for the
site. 'Articles' consist of a header with a title and an article
description, and a content portion.